requirements.txt
ultralytics
opencv-python
numpy
Pillow
tqdm
streamlit
torch
seaborn
matplotlib